feat: 회원탈퇴 soft/delete 분기 추가 및 hard delete 로직 구현

+++ nmnb-application/src/main/kotlin/nmnb/application/global/auth/service/AuthServiceImpl.kt
@@ -1,5 +1,6 @@
 package nmnb.application.global.auth.service
 
+import jakarta.persistence.EntityManager
 import nmnb.application.global.auth.service.dto.response.AuthTokenResponse
 import nmnb.application.global.auth.service.dto.response.AuthUserResponse
 import nmnb.application.global.common.utils.DeviceIdUtils
@@ -9,6 +10,7 @@ import nmnb.application.global.infrastructure.security.JwtProvider
 import nmnb.common.properties.S3Properties
 import nmnb.domain.auth.SocialType
 import nmnb.domain.user.User
+import nmnb.domain.user.WithdrawType
 import nmnb.domain.user.repository.UserRepository
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
@@ -23,6 +25,7 @@ class AuthServiceImpl(
     private val s3Properties: S3Properties,
     private val refreshTokenService: RefreshTokenService,
     private val blacklistService: BlacklistService,
+    private val entityManager: EntityManager,
 ) : AuthService {
 
     @Transactional
@@ -80,7 +83,36 @@ class AuthServiceImpl(
     }
 
     @Transactional
-    override fun withdraw(user: User) {
-        userRepository.delete(user)
+    override fun withdraw(user: User, withdrawType: WithdrawType) {
+        when (withdrawType) {
+            WithdrawType.HARD -> {
+                hardDeleteUser(user.id!!)
+            }
+            WithdrawType.SOFT -> {
+                userRepository.delete(user)
+            }
+        }
+    }
+
+    private fun hardDeleteUser(userId: String) {
+        // 유저가 누른 좋아요 삭제
+        entityManager.createNativeQuery("DELETE FROM user_post_likes WHERE user_id = :id")
+            .setParameter("id", userId)
+            .executeUpdate()
+
+        // 유저가 작성한 게시글에 달린 좋아요 삭제
+        entityManager.createNativeQuery("DELETE FROM user_post_likes WHERE post_id IN (SELECT post_id FROM posts WHERE user_id = :id)")
+            .setParameter("id", userId)
+            .executeUpdate()
+
+        // 유저가 작성한 게시글 삭제
+        entityManager.createNativeQuery("DELETE FROM posts WHERE user_id = :id")
+            .setParameter("id", userId)
+            .executeUpdate()
+
+        // 유저 삭제
+        entityManager.createNativeQuery("DELETE FROM users WHERE user_id = :id")
+            .setParameter("id", userId)
+            .executeUpdate()
     }
 }
